moved getTarget into a header file
